Updating role

+++ app/Services/Project/ProjectService.php
@@ -4,6 +4,7 @@
 namespace App\Services\Project;
 
 use App\Models\Project;
+use Illuminate\Support\Facades\Gate;
 
 class ProjectService
 {
@@ -11,7 +12,6 @@ class ProjectService
     {
         try {
             // $projects = Project::where('owner_id', $user->id)->latest()->get();
-
             $projects = $user->projects()->latest()->get();
 
             return response()->json([
@@ -23,18 +23,25 @@ class ProjectService
     }
 
 
+
+
     public function createProject($projectData)
     {
         try {
-            $project = auth()->user()->projects()->create($projectData);
+            if (Gate::allows('create', Project::class)) {
+                $project = auth()->user()->projects()->create($projectData);
 
-            return response()->json([
-                'project' => $project
-            ]);
+                return response()->json([
+                    'project' => $project
+                ]);
+            }
+            return response()->json('Unauthorized for this action');
         } catch (\Exception $e) {
             return response()->json($e->getMessage());
         }
     }
+
+
 
     public function getProjectById($projectId)
     {
@@ -50,14 +57,18 @@ class ProjectService
     }
 
 
+
+
     public function deleteProject($projectId)
     {
         try {
             $project = Project::findOrFail($projectId);
 
-            $project->delete();
-
-            return response()->json([], 200);
+            if (Gate::allows('delete', $project)) {
+                $project->delete();
+                return response()->json([], 200);
+            }
+            return response()->json('Unauthorized for this action');
         } catch (\Exception $e) {
             return response()->json($e->getMessage());
         }
